pyproject.toml
[project]
name = "vikunja-to-gcal"
version = "1.0.0"
description = "Fetches pending tasks from a Vikunja instance and schedules them as sequential events in Google Calendar, starting from the current time."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-api-python-client>=2.194.0",
    "google-auth-oauthlib>=1.3.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
